requirements.txt
langchain
langchain-groq
python-dotenv
fastapi
uvicorn
pymongo
